+++ clients/termit-client/src/index.ts
@@ -49,3 +49,4 @@ export * from "./desktopOps";
 export * from "./opsAutomation";
 export * from "./mediaOps";
 export * from "./feedbackOps";
+export * from "./orchestrationOps";